requirements.txt
<<<<<<< HEAD
python-dotenv==1.0.0
ruamel.yaml==0.17.21
websockets==10.4
requests==2.31.0
solana==0.22.0
httpx==0.18.2
pytest==7.4.0
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.9.1
typing-extensions==3.10.0.2
numpy==2.3.1
=======
solana==0.36.7
solders==0.26.0
httpx
python-dotenv
fastapi
uvicorn[standard]
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.3
>>>>>>> 2553739 (Ajout de la config Railway et du workflow CI/CD)
